でかくなってきたDict.swiftからFileDict, MemoryDictを別ファイルに分離
